# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anki Card Generator
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all decks
 */
export const ListDecksResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  parentId: zod.number().nullish(),
  cardCount: zod.number(),
  createdAt: zod.string(),
});
export const ListDecksResponse = zod.array(ListDecksResponseItem);

/**
 * @summary Create a new deck
 */
export const CreateDeckBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  parentId: zod.number().nullish(),
});

/**
 * @summary Get a deck by ID
 */
export const GetDeckParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeckResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  parentId: zod.number().nullish(),
  cardCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a deck
 */
export const UpdateDeckParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDeckBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  parentId: zod.number().nullish(),
});

export const UpdateDeckResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  parentId: zod.number().nullish(),
  cardCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a deck
 */
export const DeleteDeckParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List cards for a deck
 */
export const ListDeckCardsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListDeckCardsResponseItem = zod.object({
  id: zod.number(),
  deckId: zod.number(),
  front: zod.string(),
  back: zod.string(),
  tags: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListDeckCardsResponse = zod.array(ListDeckCardsResponseItem);

/**
 * @summary Update a card
 */
export const UpdateCardParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCardBody = zod.object({
  front: zod.string().optional(),
  back: zod.string().optional(),
  tags: zod.string().nullish(),
});

export const UpdateCardResponse = zod.object({
  id: zod.number(),
  deckId: zod.number(),
  front: zod.string(),
  back: zod.string(),
  tags: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a card
 */
export const DeleteCardParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate Anki cards from text content
 */
export const GenerateCardsBody = zod.object({
  text: zod.string(),
  deckName: zod.string(),
  cardCount: zod.number().optional(),
  parentId: zod.number().nullish(),
});

/**
 * @summary Export deck as CSV text for Anki import
 */
export const ExportDeckParams = zod.object({
  id: zod.coerce.number(),
});

export const ExportDeckResponse = zod.object({
  deckName: zod.string(),
  csv: zod.string(),
  cardCount: zod.number(),
});